chore: move supportsCookies to featureDetection
