+++ lib/utils/__tests__/featureDetection.node.test.ts
@@ -1,0 +1,13 @@
+/**
+ * @jest-environment node
+ */
+
+import { supportsCookies } from "../featureDetection";
+
+describe("featureDetection in node env", () => {
+  describe("supportsCookies", () => {
+    it("returns false", () => {
+      expect(supportsCookies()).toBe(false);
+    });
+  });
+});
