+++ lib/utils/__tests__/featureDetection.test.ts
@@ -1,0 +1,9 @@
+import { supportsCookies } from "../featureDetection";
+
+describe("featureDetection", () => {
+  describe("supportsCookies", () => {
+    it("returns true in jsdom env", () => {
+      expect(supportsCookies()).toBe(true);
+    });
+  });
+});
