--- lib/utils/__tests__/index.node.test.ts
@@ -1,11 +1,0 @@
-/**
- * @jest-environment node
- */
-
-import { supportsCookies } from "../index";
-
-describe("supportsCookies in node env", () => {
-  it("returns false", () => {
-    expect(supportsCookies()).toBe(false);
-  });
-});
