+++ lib/utils/__tests__/index.test.ts
@@ -1,10 +1,4 @@
-import {
-  isFunction,
-  isNumber,
-  isString,
-  isUndefined,
-  supportsCookies
-} from "../index";
+import { isFunction, isNumber, isString, isUndefined } from "../index";
 
 describe("isUndefined", () => {
   test.each`
@@ -51,11 +45,5 @@ describe("isFunction", () => {
     ${() => null} | ${true}
   `("should return $expected when passed $input", ({ input, expected }) => {
     expect(isFunction(input)).toEqual(expected);
-  });
-});
-
-describe("supportsCookies", () => {
-  it("returns true in jsdom env", () => {
-    expect(supportsCookies()).toBe(true);
   });
 });
