+++ lib/utils/featureDetection.ts
@@ -1,0 +1,7 @@
+export const supportsCookies = () => {
+  try {
+    return Boolean(navigator.cookieEnabled);
+  } catch (e) {
+    return false;
+  }
+};
