+++ lib/utils/index.ts
@@ -3,11 +3,4 @@ export const isUndefined = value => typeof value === "undefined";
 export const isString = value => typeof value === "string";
 export const isNumber = value => typeof value === "number";
 export const isFunction = value => typeof value === "function";
-
-export const supportsCookies = () => {
-  try {
-    return Boolean(navigator.cookieEnabled);
-  } catch (e) {
-    return false;
-  }
-};
+export * from "./featureDetection";
